ponderada2/backend/requirements.txt
fastapi==0.89.1
ormar==0.12.1
psycopg2-binary==2.9.5
uvicorn==0.20.0
asyncpg==0.28.0
sqlalchemy
passlib
pyjwt
jose
pydantic[email]
